requirements.txt
python-3.6
jupyter notebook
keras
nltk
pandas
sklear
matplotlib